Split TableName and TableNameParams into their own files
