+++ client/app/forms/form-host/form-host.component.ts
@@ -18,7 +18,7 @@ import {
     MasterTableName, TableHeader, TableMeta
 } from '../../common/api';
 import { DATE_FORMAT, DATETIME_FORMAT } from '../../common/constants';
-import { TableName } from '../../common/table-name';
+import { TableName } from '../../common/table-name.class';
 import { unflattenTableNames } from '../../common/util';
 import { TableService } from '../../core/table.service';
 import { PartialFormComponent } from '../partial-form/partial-form.component';
