+++ client/app/forms/partial-form/partial-form.component.ts
@@ -12,7 +12,7 @@ import { Subscription } from 'rxjs/Subscription';
 import * as _ from 'lodash';
 
 import { TableMeta } from '../../common/api';
-import { TableName } from '../../common/table-name';
+import { TableName } from '../../common/table-name.class';
 import { TableService } from '../../core/table.service';
 import { FormControlSpec } from '../form-control-spec.interface';
 import { FormSpecGeneratorService } from '../form-spec-generator/form-spec-generator.service';
