+++ client/app/tables/datatable.component.spec.ts
@@ -10,7 +10,7 @@ import { Observable } from 'rxjs/Observable';
 import * as sinon from 'sinon';
 
 import { SqlRow, TableMeta } from '../common/api';
-import { TableName } from '../common/table-name';
+import { TableName } from '../common/table-name.class';
 import { CoreModule } from '../core/core.module';
 import { TableService } from '../core/table.service';
 import { DatatableComponent } from './datatable.component';
