+++ common/table-name-params.interface.ts
@@ -1,0 +1,8 @@
+import { TableTier } from './api';
+
+export interface TableNameParams {
+    rawName: string;
+    tier: TableTier;
+    cleanName: string;
+    masterRawName: string | null;
+}
