+++ server/test/util.test.ts
@@ -3,7 +3,7 @@ import * as _ from 'lodash';
 import { expect } from 'chai';
 
 import { MasterTableName } from '../../common/api';
-import { TableName } from '../../common/table-name';
+import { TableName } from '../../common/table-name.class';
 import { unflattenTableNames } from '../../common/util';
 
 describe('common/util', () => {
